pyproject.toml
[project]
name = "banking-transaction-analyzer"
package-mode = false
version = "0.1.0"
description = "backend module for analyzing banking transactions and generating financial analytics"
authors = [
    {name = "Nadezhda_Popova",email = "nadezhdapopova13@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas-stubs (>=2.2.3.250527,<3.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.16.1"
isort = "^6.0.1"
black = "^25.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[tool.mypy]
disallow_untyped_defs=true
warn_return_any=true
exclude = """
/(
    .git
    .venv
    venv
)/
"""

[tool.black]
line-length = 119
exclude = """
/(
    .git
    .venv
    venv
)/
"""

[tool.isort]
line_length = 119
multi_line_output = 3